feat(settings): manage SSH connections from a dedicated settings tab
Extract the list/editor UI out of SshConnectionsModal into a shared
SshConnectionsManager component and mount it in two places: the existing
modal (connection-creation flow, unchanged behavior) and a new SSH tab in
Settings. Persistence logic stays in utils/ssh.ts.

The delete confirmation now warns how many saved database connections
reference the tunnel being removed (best-effort count via
get_connections). Typography and paddings follow the settings-tab
conventions in both contexts.

New i18n keys (manageDesc, deleteInUse with per-language plural forms)
in all 8 locales.
