+++ src/i18n/locales/de.json
@@ -750,7 +750,10 @@
     "keyFile": "Schlüsseldatei",
     "quickTest": "Schneller Verbindungstest",
     "testFailed": "Verbindungstest fehlgeschlagen",
-    "savedInKeychain": "Passwort im System-Keychain gespeichert"
+    "savedInKeychain": "Passwort im System-Keychain gespeichert",
+    "manageDesc": "Wiederverwendbare SSH-Tunnel für deine Datenbankverbindungen. Passwörter und Passphrasen werden im System-Schlüsselbund gespeichert.",
+    "deleteInUse_one": "Sie wird von {{count}} gespeicherten Verbindung verwendet, die sich nicht mehr über diesen Tunnel verbinden wird.",
+    "deleteInUse_other": "Sie wird von {{count}} gespeicherten Verbindungen verwendet, die sich nicht mehr über diesen Tunnel verbinden werden."
   },
   "sshAskpass": {
     "title": "SSH-Authentifizierung",
