+++ src/i18n/locales/it.json
@@ -755,7 +755,10 @@
     "keyFile": "File chiave",
     "quickTest": "Test connessione rapido",
     "testFailed": "Test connessione fallito",
-    "savedInKeychain": "Password salvata nel portachiavi di sistema"
+    "savedInKeychain": "Password salvata nel portachiavi di sistema",
+    "manageDesc": "Tunnel SSH riutilizzabili per le tue connessioni ai database. Password e passphrase sono salvate nel portachiavi di sistema.",
+    "deleteInUse_one": "È usata da {{count}} connessione salvata, che non passerà più da questo tunnel.",
+    "deleteInUse_other": "È usata da {{count}} connessioni salvate, che non passeranno più da questo tunnel."
   },
   "sshAskpass": {
     "title": "Autenticazione SSH",
