+++ src/i18n/locales/ja.json
@@ -764,7 +764,9 @@
     "keyFile": "鍵ファイル",
     "quickTest": "簡易接続テスト",
     "testFailed": "接続テストに失敗しました",
-    "savedInKeychain": "パスワードをシステムのキーチェーンに保存しました"
+    "savedInKeychain": "パスワードをシステムのキーチェーンに保存しました",
+    "manageDesc": "データベース接続で再利用できるSSHトンネルです。パスワードとパスフレーズはシステムのキーチェーンに保存されます。",
+    "deleteInUse_other": "この接続は{{count}}件の保存済みデータベース接続で使用されており、削除するとこのトンネル経由で接続できなくなります。"
   },
   "sshAskpass": {
     "title": "SSH認証",
