+++ src/i18n/locales/ru.json
@@ -743,7 +743,12 @@
     "keyFile": "Файл ключа",
     "quickTest": "Быстрая проверка подключения",
     "testFailed": "Проверка подключения не удалась",
-    "savedInKeychain": "Пароль сохранён в системном хранилище ключей"
+    "savedInKeychain": "Пароль сохранён в системном хранилище ключей",
+    "manageDesc": "Многоразовые SSH-туннели для подключений к базам данных. Пароли и парольные фразы хранятся в системной связке ключей.",
+    "deleteInUse_one": "Его использует {{count}} сохранённое подключение, которое больше не будет подключаться через этот туннель.",
+    "deleteInUse_few": "Его используют {{count}} сохранённых подключения, которые больше не будут подключаться через этот туннель.",
+    "deleteInUse_many": "Его используют {{count}} сохранённых подключений, которые больше не будут подключаться через этот туннель.",
+    "deleteInUse_other": "Его используют {{count}} сохранённых подключений, которые больше не будут подключаться через этот туннель."
   },
   "sshAskpass": {
     "title": "Аутентификация SSH",
